requirements.txt
# python==3.6.4
jinja2==2.10
Pillow==5.0.0
matplotlib==2.1.2
werkzeug==0.14.1
keras==2.1.3
tensorflow==1.5.0
imgaug==0.2.5
opencv-python==3.4.0.12
falcon==1.4.1
numpy==1.14.0
h5py==2.7.1
gunicorn==19.7.1
git+https://github.com/broadinstitute/keras-resnet
git+https://github.com/fizyr/keras-retinanet.git@0fe9fe66583c414d83612748305d7b430902376
openslide-python
shapely
pydash